Add resource presets
- Mods can now provide default presets using resource packs
- Add server and resource saved hotkey groups
- Mods can now provide default saved hotkey groups, and default groups that will be loaded for the first time, using resource packs
- Unknown hotkey actions from unloaded mods are now preserved
- Added network guards to check if the server has the mod installed
- Redesign server config permissions to allow view-only access and separate preset saving permissions.

+++ src/main/java/endorh/simpleconfig/core/entry/EntryPairEntry.java
@@ -64,12 +64,12 @@ public class EntryPairEntry<
 		}
 		
 		public Builder<L, R, LC, RC, LG, RG, LE, RE, LB, RB> withSplitPosition(
-		  @Range(from = 0, to = 1) float splitPosition
+		  @Range(from = 0, to = 1) double splitPosition
 		) {
 			if (splitPosition < 0 || splitPosition > 1) throw new IllegalArgumentException(
 			  "Split position must be between 0~1. Specifically: " + splitPosition);
 			final Builder<L, R, LC, RC, LG, RG, LE, RE, LB, RB> copy = copy();
-			copy.splitPos = splitPosition;
+			copy.splitPos = (float) splitPosition;
 			return copy;
 		}
